src/recognition_api/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7

tensorflow==1.14.*
Keras==2.2.4

keras-vggface

# torch==1.4.0              
# torchvision==0.5.0
torch==1.5.1   # as per retinaface requirements           
torchvision==0.6.1

mtcnn==0.0.9
opencv-python==4.2.0.*

# retina face detector
fdet==0.2.1

Pillow==6.2.1
imageio==2.6.1

sentry-sdk[flask]==0.14.3
